1. update Conv_Coulomb_Pot_K parameter 	1.1 src_lcao/conv_coulomb_pot_k.h 	1.2 src_lcao/conv_coulomb_pot_k.cpp 	1.3 src_lcao/conv_coulomb_pot_k-template.h 	1.4 src_lcao/exx_lcao.cpp 2. Conv_Coulomb_Pot_K output rmesh_times 	2.1 src_lcao/conv_coulomb_pot_k.h 	2.2 src_lcao/conv_coulomb_pot_k.cpp 	2.3 src_lcao/conv_coulomb_pot_k-template.h 	2.4 src_lcao/exx_lcao.h 	2.5 src_lcao/exx_lcao.cpp 	2.6 src_lcao/abfs.h 	2.7 src_lcao/abfs.cpp

+++ ABACUS.1.0.0/source/src_lcao/abfs.cpp
@@ -148,7 +148,7 @@ map<size_t,map<size_t,map<Abfs::Vector3_Order<int>,shared_ptr<matrix>>>> Abfs::c
 	const vector<Vector3_Order<int>> &Coulomb_potential_boxes,
 	const Exx_Abfs::Matrix_Orbs11 &m_abfs_abfs,
 	const Element_Basis_Index::IndexLNM &index_abfs,
-	const int rmesh_times,
+	const double rmesh_times,
 	const double threshold,
 	map<size_t,map<size_t,map<Vector3_Order<double>,weak_ptr<matrix>>>> &Vws )
 {
